+++ lambda/shared/utils/email.ts
@@ -1,0 +1,4 @@
+export const isEmail = (email: string): boolean => {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    return emailRegex.test(email);
+}
