+++ lambda/shared/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './http-response'
+export * from './email'
